Fix building of Harfbuzz
Building of Harfbuzz was failing when ramses-sdk_ALLOW_PLATFORM_FREETYPE
is set to OFF and a Freetype libraray >= 2.7.1 is already installed in
the platform.

Hafbuzz CMakeLists.txt uses CMake's check_function_exists() to check
for the existence of this Freetype functions:

FT_Get_Var_Blend_Coordinates
FT_Set_Var_Blend_Coordinates
FT_Done_MM_Var

This functions are available since Freetype version 2.7.1. The
delivered Freetype version that comes with RAMSES is 2.6, which does
not have this functions.

Now, when a Freetype libaray >= 2.7.1 is installed in the platform,
check_function_exists() will return TRUE for the functions, and
the build will fail later in hb-ft.cc when the functions are tried
to be called, but are not declared in the Freetype version from RAMSES.