Oss release 2019-6-14
New features:
- added support for scissoring on client API
- first implementation of DCSM (display cluster state management) - see
   below for more info
Cosmetic API changes:
- make Appearance::set/getDepthWrite() API symmetric (both use enums
    instead of bool now)
- getStatusMessage() moved to StatusObject for consistency
Improvements:
- made DisplayManager utility class more stable wrt. to scene states and
    handling

DCSM:
The ramses API so far only provides means to create and stream 3D scenes.
It does not provide a way to control which scene is shown on which
renderer on a distributed system. The new DCSM service(s) are supposed to
make it easier to implement software stacks which can also control and
synchronize at which point a scene is created, and negotiate the state of
the scene - should it be kept up-to-date, is it about to disappear, at
which point should it start its animations and so on. The feature is not
fully documented yet, except the API itself - documentation and overview
coming in future releases.

Original commit sha: e381bba923be0c1feed737a4299f0ffdfeb8c388