Add alternate constructor for WaylandEGLExtensionProcs
Add a second constructor which accepts a pre-configured
EGLDisplay rather than requiring the user to supply the
wl_display connected to the system compositor.

This will be used in upcoming changes that need to instantiate
the WaylandEGLExtensionProcs utility class from down inside
the implementation of a nested compositor. At that point, it
is difficult to retrieve the wl_display which is held as a
private detail of the rendering backend.